Handle errors and retry requests
The design of the error handling is based on having a MessageBuffer hold
messages for all topics/partitions. Whenever we want to flush messages
to the cluster, we group the buffered messages by the broker they need
to be sent to and fire off a request to each broker. A request can be a
partial success, so we go through the response and inspect the error
code for each partition that we wrote to. If the write to a gived
partition was successful, we clear the corresponding messages from the
buffer -- otherwise, we log the error and keep the messages in the
buffer.

After this, we check if the buffer is empty. If it is, we're all done.
If it's not, we do another round of requests, this time with just the
remaining messages. We do this for as long as `max_retries` permits.