Skip responses to previous requests
There may have been a previous request that timed out before the client
was able to read the response. In that case, the response will still be
sitting in the socket waiting to be read. If the response we just read
was to a previous request, we can safely skip it.