Allow buffering messages even when Kafka is unavailable
Until now, `Producer#produce` has required access to the cluster
metadata in order to determine the partition that a given message should
be assigned to. In the case of Kafka being unavailable, this would have
blocked clients from buffering messages, as the metadata request would
fail. Furthermore, since ffc1a34 the metadata has been invalidated the
first time `#produce` is called with a topic not seen by the Producer
instance before. This would mean a thundering herd of metadata requests
when e.g. a cluster of web application processes are restarted. By
postponing the metadata request until we're ready to send all the
messages, we'll be able to do a single request containing *all* the
topics.