Avoid a barrage of metadata requests on errors
Previously, `BrokerPool#mark_as_stale!` would cause the cached metadata
to be removed, but each subsequent call for metadata info (such as
leader lookups) would cause a metadata request. In order to avoid this
situation, the metadata is kept in the cache even though it's stale.
This makes sense because the metadata is not either completely accurate
or completely inaccurate -- it can be accurate for some partitions but
not others. Therefore we get as much work done as possible with the
existing metadata and only do a refresh when we start a new retry.