Set a no output logger by default
There are a lot of calls to `logger` so either a nil logger argument
needs to either raise an ArgumentError or have a reasonable default.

Here I'm presuming that the logger argument being nil is merely
development exploration of the gem, so defaulting to a nil logger with a
warning is acceptable.

Fixes #121