Allow configuring how producer buffer overflows are handled
By default, an exception will be raised if the producer's buffer becomes
full. This change will allow the client to configure the producer to
instead prune the oldest messages in the buffer but keep the message
that's currently being written. This handles an error condition where a
topic or partition has is permanently unavailable for writes, making
messages destined for it fill up the buffer and preventing messages for
other topics from being delivered. By using this policy, the messages
least likely to ever be successfully delivered -- those that have been
buffered for the longest time -- will be dropped, rather than new
messages that may have a chance of being delivered.