Add a simple API for synchronously sending a single message
This is simple API that offers neither performance or fault tolerance,
making it ideal for Ruby! More seriously, it addresses a very specific
scenario: sending one-off messages. Currently, you have to be careful to
avoid leaving messages in a Producer buffer when something goes wrong,
and setting up a producer in the first place when all you want is
something hammer-shaped that sends a message to Kafka is something you'd
expect from a Java library.