Eagerly resolve default offsets in the consumer
Previously, the first time a partition was consumed there would be no
committed offset, so the consumer would use the default offset (either
the earliest or the latest offset in the partition.) However, the actual
offset number would only be resolved when the consumer fetched from the
partition. This would leave a race condition when a consumer subscribes
to the latest messages for a topic: if no messages arrive while the
fetch is waiting, no offset will be committed; in the next iteration,
the offset will be resolved anew, but that means skipping any message
that has arrived in the meantime. Only when messages arrive *while* the
consumer is doing a fetch on a partition will this race be broken.

The new behavior is to resolve the default offset as soon as possible
and to commit the actual offset. This means that after a consumer has
started the starting offsets are fixed.