Enable keepalive for the sockets
The Kafka brokers automatically kill client connections after a period
of inactivity, typically 10 minutes. In order to avoid getting a network
error and having to re-connect, the sockets should be kept alive.

Unfortunately, Linux and Mac OS are not consistent in the APIs they
offer, so we have to use both.