Don't discard all consumer offset data on resync
When a group resyncs we were discarding all local data about which
partition offsets had been processed. Any data that had not been
committed would be lost, even if the consumer was assigned some of the
same partitions after the sync. With this change, consumers only discard
offset data for partitions they are no longer assigned.

This should result in less reprocessing of messages.