Fix invalid BufferOverflow error in AsyncProducer
The shutdown method used to terminate only the Worker thread and not
the Timer thread. The Timer thread continues to push `deliver_messages`
to the queue but the Worker thread is not alive to consume them.
After some time the queue fills with `deliver_messages` and the next
try to produce a message fails with Kafka::BufferOverflow.

This commit fixes the issue by terminating the Timer thread
together with the Worker thread on shutdown.