fix ssl socket timeout
We've been seeing a case where Kafka won't accept new SSL connections
properly. It accepts the TCP one, but doesn't perform the SSL handshake
properly. The SSL socket behaviour in this library doesn't handle this
case properly.

This clears up the logic in this class as well to use a
`select_with_timeout` method (taken from
https://github.com/excon/excon/blob/192bb842e70b550ed6d78c6d621962ee0faca4be/lib/excon/socket.rb)