Add host and port to BrokerPool cache key
In cases where a Kafka broker goes down and new broker with a different host/port spins up with the same broker id, we don't want to use the previously cached broker. To fix this, we'll consider the host and port when deciding whether or not to use the cached broker. If the cached broker is no longer valid, we'll disconnect and close the connection.