Raise MessageTooLargeToRead if the broker has truncated the message set
When the client specifies a maximum number of bytes to read from a
partition and the first message in the partition exceeds that, the
broker will truncate the message set to match the client's maximum. This
means that the client is unable to decode any messages.

Rather than infinitely keep asking for messages just to have a truncated
message returned, we not raise MessageTooLargeToRead.