Stop fetcher thread synchronously
Stopping the fetcher asynchronously (by emitting a message onto a queue
for pickup) results in a potential race condition on consumer processing
errors, wherein the consumer can be restarted while the fetcher is
sleeping, resulting in the fetcher being stopped after the consumer has
already started. This results in processing errors taking down the
consumer in a hard-to-detect way: the fetcher thread, being stopped, no
longer fetches any new messages, but the consumer continues to loop as
though everything is fine.

Fix this by waiting until the fetcher thread to stops when it's asked
to. I don't think there are performance implications here: given that we
only stop the fetcher before committing offsets and leaving the consumer
group, there doesn't appear to be any benefit to leaving it running.