Verify SSL Certificates
The default in Ruby is to verify the certificate, but then ignore the
result of that verification.
So it needs to be a structurally valid cert, but it's unrelated to the
connection being opened or the CAs we've decided to trust.

This enables verification so we'll only trust certs we should trust.
It also sets the hostname so the cert presented has to correspond to the
host we're connecting to.