Correctly handle timestamp deltas as milliseconds
Kafka timestamps are always given in milliseconds, which is why
first_timestamp was already divided by 1000 here. However, the
timestamp_delta was then added as is, which is incorrect -- adding
milliseconds to a timestamp in seconds will always produce the
wrong result (except in the degenerate case where the delta is
zero).

We could simply divide the delta by 1000 in the same way that was
already being done for the first timestamp, but that would discard
the sub-second part of the timestamp. Instead, divide each
timestamp by BigDecimal(1000), to preserve full precision while
avoiding floating point rounding errors.

Of course, we also need to multiply the delta by 1000 when
calculating it so that produced messages are correct as well.

With this patch, the record timestamps reported by ruby-kafka match
those from <https://github.com/fgeller/kt>, so I'm reasonably
confident this is correct.

I believe this is a fix for
<https://github.com/zendesk/ruby-kafka/issues/721>.