Send JoinGroupRequest v1 messages with rebalance timeouts
If we send a JoinGroupRequest v0 message to join an existing
consumer group, the broker does not wait for other members to send
their own JoinGroupRequests, resulting in all other members being
kicked out of the group. When they rejoin later, the same thing
happens, resulting in the group continually being rebalanced with
a single member each time.

Sending a JoinGroupRequest v1 message instead lets us specify a
timeout to wait for other members. A new member is now added to
the consumer group as expected, rather than becoming the only
member.

The default value of 1 minute for rebalance_timeout is taken from
what Sarama (Go library) does. librdkafka defaults to 5 minutes
here, which seems a little excessive.

This will break if talking to a Kafka broker older than 0.10.1.
This is unlikely to be a problem because these messages are only
used in ruby-kafka to support consumer groups, which were already
broken due to the issue described above.

While here, also send the correct group protocol information -- we
only support round-robin assignment, and we have an actual list of
topics rather than "test-messages". This is needed for
interoperability with other clients participating in the same
consumer group (found when debugging this issue with Sarama).