Have a uniform expectation for producing methods about topic's type
This commit was written because some of my users (I maintain a library that
wraps ruby-kafka) were producing asynchronously with the topic argument being a
Symbol, and not seeing any problems until messages couldn't be added to the
async message queue because the thread producer couldn't deliver messages with
Symbol topics.

However, we can't just ensure .to_s is called, since every Ruby object has a
String representation. So we ensure that `topic` is either a Symbol or a String.

I considered limiting it to Strings, but if a user provides a symbol `topic`
argument, it's clear what they want (as opposed to providing, say, a Hash), and
it doesn't seem harmful to allow that.

The reason I updated all 3 producing methods is that it seems nicer to have an
explanatory error message upfront instead getting a NoMethodError for a method
named `bytesize`.