Recover from Kafka::MessageSizeTooLarge exception in async producer.
You can get into a bad unrecoverable state when a Kafka::MessageSizeTooLarge exception is raised.

When we receive this exception, if we're using the sync producer, then we simply re-raise this
exception. If we're using the async producer, then we attempt to remove the largest message from
the buffer and simply retry. Once retried, if the buffer is still too large, we'll continue removing
the largest message until we can successfully flush the buffer.