Return early in `ensure_threads_running!` if there's nothing to do
We don't need to grab the mutex if both threads are still alive. But if
we _do_ need to grab the mutex (i.e., if either/both of the threads are
no longer alive), we'll need to re-check that the threads are alive
later since someone else may have already restarted them by the time
we've grabbed the mutex.