It's valid to abort a txn that hasn't been started
There may be cases where dependant code starts a transaction using the
producer's `#transaction` block syntax and then an exception is raised
within that block *before* the transaction manager actually opens up a
transaction. In such cases, prior to this change, the real exception is
obfuscated, because the transaction manager would raise an exception
about not being in a valid transactional state.

With this change, attempting to abort a transaction that has not
actually been opened will no longer result in an exception. Instead it
will log a warning message and leave the transaction manager in its
ready state.