Make example "create partitions" stable
As there is a time lag for Kafka's metadata to be updated,
Kafka::Client#partitions_for is needed to be executed until the
metadata is updated to check the latest partition count.
However, Kafka::Cluster never fetches new metadata until it
considers the current metadata stale.
This commit adds the method Kafka#refresh_metadata in order to
force Kafka::Cluster to fetch metadata. The option helps in
production code as well to ensure we get the latest metadata.